Remove pc chart, postprocess records, make blob download available
The PC chart is removed, as records being linked to specific characters could
never be released publically.

The records are being postprocessed, as information about stats for unknown
records could have been logged after the record (ie. STAT on LOGOUT)

The processed records are available for download in JSON format. Currently,
this is only accessible by calling parser.createBlobLink() from the console.
The process can be quite resource intensive.
The file should be saved with "scrubbed" appearing somewhere in the filename.

The loader now properly informs its each method of the current file and the
number of files.

Files with "scrubbed" in the filename are interpreted as scrubbed logfiles and
will be loaded in as JSON directly.