Compatability fixes
This is by no means even close to comprehensive (legacy compatability for
determining browser width is a total mess in and of itself).

However, this is enough for manavis to work on recent versions of Firefox.
It's not quite as fast as chromium, but it is entirely workable.

Older versions aren't really worth testing, as the entire application is very
js intensive.